Move account_invoice from carrier_info to validate_send_email module
