+++ stock_carrier_info/models/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import account_invoice
 from . import stock_picking
 from . import stock_carrier_info
