--- stock_carrier_info/views/account_invoice_view.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="account_invoice_auto_dispatch_inherit" model="ir.ui.view">
-        <field name="name">account.invoic.carrier.info.inherit</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='move_id']" position="after">
-                <field name="carrier_info_name" />
-                <field name="carrier_info_url" />
-            </xpath>
-        </field>
-    </record>
-</odoo>
